KCIRB-265 : Add authorization check for delete protocol/amendment/renewal

+++ src/main/java/org/kuali/kra/irb/actions/ProtocolProtocolActionsAction.java
@@ -444,8 +444,10 @@ public class ProtocolProtocolActionsAction extends ProtocolAction implements Aud
             HttpServletResponse response) throws Exception {
         
         ProtocolForm protocolForm = (ProtocolForm) form;
-        getProtocolDeleteService().delete(protocolForm.getProtocolDocument().getProtocol(), protocolForm.getActionHelper().getProtocolDeleteBean());
-
+        ProtocolTask task = new ProtocolTask(TaskName.PROTOCOL_AMEND_RENEW_DELETE, protocolForm.getProtocolDocument().getProtocol());
+        if (isAuthorized(task)) {
+           getProtocolDeleteService().delete(protocolForm.getProtocolDocument().getProtocol(), protocolForm.getActionHelper().getProtocolDeleteBean());
+        }
         return mapping.findForward(MAPPING_BASIC);
     }
